new header directory, no loginbutton on login page
