+++ src/components/Root.tsx
@@ -1,5 +1,5 @@
 import React, { useContext, useEffect } from "react";
-import { Header } from "./Header";
+import { Header } from "./header/Header";
 import { useNavigate, Outlet } from "react-router-dom";
 import { AuthContext } from "../contexts/AuthProvider";
 import { LoginPage } from "../pages/login/LoginPage";
